b/f: avoid to hide AmAudio::close() with AmPlaylist::close(…)
- AmPlaylist::close(…) is now replaced by AmPlaylist::flush().
- AmAudio::noAudio event is only sent if the play-cursor reaches the end (not through flush() or close()).

+++ apps/examples/tutorial/cc_acc/CCAcc.cpp
@@ -108,7 +108,7 @@ int CCAcc::subtractCredit(string pin, int amount) {
     credits_mut.unlock();
     return -1;
   }
-  if (it->second - amount < 0) {
+  if ((int)it->second - amount < 0) {
     credits[pin] = 0;
   } else {
     credits[pin] = it->second - amount;
